+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- SPFarm
+  - Added parameter SkipRegisterAsDistributedCacheHost
+
 ## [4.7.0] - 2021-06-10
 
 ### Added